surferosa-api/requirements.txt
flask==1.0.2
flask_restful==0.3.6
flask_script==2.0.6
flask_migrate==2.2.1
flask_marshmallow==0.9.0
pymongo==3.7.1
flask_pymongo==2.1.0
jsonschema==2.5.1
flask-jwt-extended==3.12.1
passlib==1.7.1